feat: 프론트 이벤트 페이지 구현
