+++ frontend/sin/src/components/event/EventDetails.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const EventDetails = () => {
+  return (
+      <div>
+        
+      </div>
+  );
+};
+
+export default EventDetails;
