+++ frontend/sin/src/components/event/FindEventsApi.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import axios from "axios";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
+
+const findEventsApi = () => {
+  return (axios.get(BACKEND_ADDRESS + "/shop/goods/event")
+  .then(response => response.data))
+};
+
+export default findEventsApi;
